refactor: colocate walk editor components under app/(editor)
